[nrf fromtree] net: sockets: tls: Free DTLS server active session on errors
In case of errors on an active session (in most cases peer closing the
session), the session should be freed. Note, that as mbed TLS needs some
session context to work with, the last session on a socket is not freed,
but only reset instead.

Signed-off-by: Robert Lubos <robert.lubos@nordicsemi.no>
(cherry picked from commit 4197720be805fe119ebcfc1f01c693b2c20079cf)